LayoutTests/js/dom/jit-set-profiling-access-type-only-for-get-by-id-self-expected.txt

Tests that we use the correct profiling accessType for the case we end up compiling/patching.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(new L_()) is 1365109051943.000000
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(new L_()) is 1365109051943.000000
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(new L_()) is 1365109051943.000000
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(new L_()) is 1365109051943.000000
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(new L_()) is 1365109051943.000000
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(new L_()) is 1365109051943.000000
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS f(129681120) is 129681120
PASS successfullyParsed is true

TEST COMPLETE
